refactor: move byte specific pieces into iroh-bytes
