+++ iroh-bytes/Cargo.toml
@@ -12,3 +12,42 @@ repository = "https://github.com/n0-computer/iroh"
 rust-version = "1.65"
 
 [dependencies]
+anyhow = { version = "1", features = ["backtrace"] }
+bao-tree = { version = "0.3.2", features = ["tokio_fsm"], default-features = false }
+blake3 = "1.3.3"
+bytes = { version = "1.4", features = ["serde"] }
+data-encoding = "2.3.3"
+derive_more = { version = "0.99.17", git = "https://github.com/JelteF/derive_more", features = ["debug", "display", "from", "try_into"] }
+flume = "0.10.14"
+futures = "0.3.25"
+hex = "0.4.3"
+iroh-net = { path = "../iroh-net" }
+multibase = "0.9.1"
+num_cpus = "1.15.0"
+once_cell = "1.17.0"
+portable-atomic = "1"
+postcard = { version = "1", default-features = false, features = ["alloc", "use-std", "experimental-derive"] }
+quic-rpc = { version = "0.6", default-features = false, features = ["quinn-transport", "flume-transport"] }
+quinn = "0.10"
+range-collections = "0.4.0"
+self_cell = "1.0.1"
+serde = { version = "1", features = ["derive"] }
+serde-error = "0.1.2"
+smallvec = { version = "1.10.0", features = ["serde", "const_new"] }
+subtle = "2.4"
+thiserror = "1"
+tokio = { version = "1", features = [] }
+tokio-stream = "0.1"
+tokio-util = { version = "0.7", features = ["io-util", "io", "rt"] }
+tracing = "0.1"
+tracing-futures = "0.2.5"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+walkdir = "2"
+
+[dev-dependencies]
+http-body = "0.4.5"
+proptest = "1.0.0"
+rand = "0.8"
+testdir = "0.7.2"
+tempfile = "3.4"
+tokio = { version = "1", features = ["macros", "test-util"] }
