+++ iroh-bytes/docs/img/get_machine.drawio
