+++ iroh-net/src/hp/magicsock/rebinding_conn.rs
@@ -224,10 +224,12 @@ mod tests {
     use super::*;
     use anyhow::Result;
 
+    const ALPN: [u8; 9] = *b"n0/test/1";
+
     fn wrap_socket(conn: impl AsyncUdpSocket) -> Result<(quinn::Endpoint, key::node::SecretKey)> {
         let key = key::node::SecretKey::generate();
         let tls_server_config =
-            tls::make_server_config(&key.clone().into(), vec![tls::P2P_ALPN.to_vec()], false)?;
+            tls::make_server_config(&key.clone().into(), vec![ALPN.to_vec()], false)?;
         let server_config = quinn::ServerConfig::with_crypto(Arc::new(tls_server_config));
         let mut quic_ep = quinn::Endpoint::new_with_abstract_socket(
             quinn::EndpointConfig::default(),
@@ -236,12 +238,8 @@ mod tests {
             Arc::new(quinn::TokioRuntime),
         )?;
 
-        let tls_client_config = tls::make_client_config(
-            &key.clone().into(),
-            None,
-            vec![tls::P2P_ALPN.to_vec()],
-            false,
-        )?;
+        let tls_client_config =
+            tls::make_client_config(&key.clone().into(), None, vec![ALPN.to_vec()], false)?;
         let client_config = quinn::ClientConfig::new(Arc::new(tls_client_config));
         quic_ep.set_default_client_config(client_config);
         Ok((quic_ep, key))
