+++ iroh-net/src/util.rs
@@ -1,10 +1,15 @@
 use std::{
     future::Future,
+    net::SocketAddr,
     pin::Pin,
+    sync::Arc,
     task::{Context, Poll},
+    time::Duration,
 };
 
 use futures::FutureExt;
+
+use crate::tls::{Keypair, PeerId};
 
 /// A join handle that owns the task it is running, and aborts it when dropped.
 #[derive(Debug)]
@@ -28,4 +33,24 @@ impl<T> Drop for AbortingJoinHandle<T> {
     fn drop(&mut self) {
         self.0.abort();
     }
+}
+
+pub fn create_quinn_client(
+    bind_addr: SocketAddr,
+    peer_id: Option<PeerId>,
+    alpn_protocols: Vec<Vec<u8>>,
+    keylog: bool,
+) -> anyhow::Result<quinn::Endpoint> {
+    let keypair = Keypair::generate();
+
+    let tls_client_config =
+        crate::tls::make_client_config(&keypair, peer_id, alpn_protocols, keylog)?;
+    let mut client_config = quinn::ClientConfig::new(Arc::new(tls_client_config));
+    let mut endpoint = quinn::Endpoint::client(bind_addr)?;
+    let mut transport_config = quinn::TransportConfig::default();
+    transport_config.keep_alive_interval(Some(Duration::from_secs(1)));
+    client_config.transport_config(Arc::new(transport_config));
+
+    endpoint.set_default_client_config(client_config);
+    Ok(endpoint)
 }
